refactor cardItemidFilterRegex.ts
rename to regular-expressions.ts
add additional regex
update imports
