+++ src/pages/import/components/settings/cardCustomization/AppCardTagSettings.tsx
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { useLazyGetItemsQuery } from '../../../../../api/codeBeamerApi';
 import { updateAppCard } from '../../../../../api/miro.api';
 import getColorForFieldLabel from '../../../../../api/utils/getColorForFieldLabel';
-import { CARD_TITLE_ID_FILTER_REGEX } from '../../../../../constants/cardTitleIdFilterRegex';
+import { CARD_TITLE_ID_FILTER_REGEX } from '../../../../../constants/regular-expressions';
 import {
 	DEFAULT_RESULT_PAGE,
 	MAX_ITEMS_PER_IMPORT,
